added input check

+++ src/main/java/sus/project/rentalSystem/controller/UserController.java
@@ -1,11 +1,14 @@
 package sus.project.rentalSystem.controller;
 
+import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.ObjectError;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -15,6 +18,7 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import jakarta.validation.Valid;
 import sus.project.rentalSystem.entity.User;
 import sus.project.rentalSystem.form.UserForm;
+import sus.project.rentalSystem.service.RentalService;
 import sus.project.rentalSystem.service.UserService;
 
 @Controller
@@ -22,6 +26,8 @@ public class UserController {
 	
 	@Autowired
 	UserService userService;
+	@Autowired
+	RentalService rentalService;
 	
 	@GetMapping("/user_list")
 	public String user_list(Model model) {
@@ -40,7 +46,10 @@ public class UserController {
 	}
 	
 	@GetMapping("/user_register")
-	public String user_register() {
+	public String user_register(Model model) {
+		if (!model.containsAttribute("userForm")) {
+	        model.addAttribute("userForm", new UserForm());
+	    }
 		return "user_register";
 	}
 	
@@ -56,9 +65,22 @@ public class UserController {
 	
 	@PostMapping("addUser")
 	public String addUser(@Valid @ModelAttribute UserForm userForm, BindingResult result, RedirectAttributes redirectAttributes) {
-		if(result.hasErrors()) {
-			System.out.println(result.getAllErrors());
-			return("redirect:user_list");
+		if(result.hasErrors()) { // Redirect to the form page
+			 List<String> errorList = result.getAllErrors().stream()
+                     .map(ObjectError::getDefaultMessage)
+                     .collect(Collectors.toList());
+			 redirectAttributes.addFlashAttribute("validationErrors", errorList);
+			 redirectAttributes.addFlashAttribute("userForm", userForm);
+			return("redirect:user_register");
+		}
+		
+		Optional<User> device = userService.findById(userForm.getEmployee_no());
+		if(device.isPresent()) {
+			//if device exists display message
+			redirectAttributes.addFlashAttribute("message", "社員番号が存在しています");
+			redirectAttributes.addFlashAttribute("alertType", "alert-warning");
+			redirectAttributes.addFlashAttribute("userForm", userForm);
+			return "redirect:user_register";
 		}
 		
 		userService.save(
@@ -101,14 +123,22 @@ public class UserController {
 			);
 			redirectAttributes.addFlashAttribute("message", "社員を編集しました");
 			redirectAttributes.addFlashAttribute("alertType", "alert-primary");
-			return("redirect:user_list");
+			return("redirect:user_info?id="+userForm.getEmployee_no());
 			}
 		return("redirect:user_list");
 	}
 	
 	@PostMapping("deleteUser")
 	public String deleteUser(@RequestParam("employee_no") String employee_no, RedirectAttributes redirectAttributes) {
+		if(rentalService.isEmployeeNoExists(employee_no)) {
+			//if in rental then refuse to delete
+			redirectAttributes.addFlashAttribute("message", "貸出中のため削除できません");
+			redirectAttributes.addFlashAttribute("alertType", "alert-danger");
+			return "redirect:/user_info?id=" + employee_no;
+		}
+		
 		Optional<User> optionalUser = userService.findById(employee_no);
+		
 		if(optionalUser.isPresent()) {
 			userService.delete(employee_no);
 			redirectAttributes.addFlashAttribute("message", "社員を削除しました");
